Decompose entity tool executor into operation-specific files
